Binary Search
Today I solved one more problem of Binary search which was about to search index. Given a sorted array of distinct integers and a target value, return the index if the target is found. If not, return the index where it would be if it were inserted in order.
It was quite easy and boost my concept of Binary Search algorithm.