Binary Search
Today I solved one more problem of binary search about search index. Given a sorted array of distinct integers and a target value, return the index if the target is found. If not, return the index where it would be if it were inserted in order
It was quite easy but it boosted my concept about Binary Search.